Fix incorrect `new` signature
It was rather a coincidence that it worked all this time due to
questionale delegation of coercions onto `Mu`. So, when a named
parameter is not specificed it defaults to `Array()` which didn't match
against `[Real, Real, ...]` only for the above mentioned reason. With
soon to be merged new coercion semantics both `Array` and `Array()`
behave similar resulting in `inconsistent bind` error.